Minor Fortran languge fixes (warnings issued by Intel's ifc)

+++ cvode/fcmix/examples_ser/cvkryf.f
@@ -28,7 +28,7 @@ C The second and third dimensions of Y here must match the values of
 C MESHX and MESHY, for consistency with the output statements below.
       DOUBLE PRECISION ATOL, AVDIM, DELT, FLOOR, RTOL, T, TOUT, TWOHR
       DOUBLE PRECISION Y(2,10,10), ROPT(40)
-      INTEGER*4 IOPT(40)
+      INTEGER IOPT(40)
       DATA TWOHR/7200.0D0/, RTOL/1.0D-5/, FLOOR/100.0D0/,
      1     JPRETYPE/1/, IGSTYPE/0/, MAXL/0/, DELT/0.0D0/
       DATA LNST/4/, LNFE/5/, LNSETUP/6/, LNNI/7/, LNCF/8/,
